parse_rr and parse_qn request a list of entries
parse_rr may be use for any kind of rr (answer, authority or
additionnal), but needs to indique which list to populate.
parse_qn is an uncompleted parse_rr and it is useful only for rr_qn
list. But to have the same arguments on the both functions, we pass the
rr_qn list like parse_rr instead the message.